+++ src/chatstyle/security/__init__.py
@@ -1,0 +1,5 @@
+"""Sensitive-value masking and secret prompt helpers."""
+
+from .mask import format_current_secret, mask_secret, prompt_sensitive_value
+
+__all__ = ["format_current_secret", "mask_secret", "prompt_sensitive_value"]